requirements.txt
NumPy==1.19.2
Pillow==8.0.0
